finish files position
