+++ client/src/views/in/career/index.js
@@ -1,2 +1,14 @@
 import Career from './Career'
-export default Career
+import { selectCareer, setKeywords, clearKeywords } from '../../../slices/careerSlice'
+import eesa from '../../../assets/images/eesa-icon.png'
+import Recruitment_image from '../../../assets/images/Recruitment.png'
+import Recommendation_image from '../../../assets/images/Recommendation.png'
+export {
+  Career,
+  selectCareer,
+  setKeywords,
+  clearKeywords,
+  eesa,
+  Recommendation_image,
+  Recruitment_image,
+}
