+++ client/src/views/in/career/recruitment/Recruitment.js
@@ -1,8 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import axios from 'axios'
-import { selectCareer, setKeywords, clearKeywords } from '../../../../slices/careerSlice'
-// import RecruPosts from './RecruPosts'
+import { selectCareer, setKeywords, clearKeywords } from '../index'
 import CareerBlock from '../CareerBlock'
 import Masonry from 'react-masonry-css'
